requirements.txt
pika==1.1.0
sqlalchemy
psycopg2
shovel
pyyaml
minio
